+++ mmseg/models/utils/__init__.py
@@ -1,5 +1,8 @@
+from .inverted_residual import InvertedResidual
 from .make_divisible import make_divisible
 from .res_layer import ResLayer
 from .self_attention_block import SelfAttentionBlock
 
-__all__ = ['ResLayer', 'SelfAttentionBlock', 'make_divisible']
+__all__ = [
+    'ResLayer', 'SelfAttentionBlock', 'make_divisible', 'InvertedResidual'
+]
