+++ mmseg/models/utils/inverted_residual.py
@@ -1,22 +1,29 @@
-from mmcv.cnn import ConvModule, build_norm_layer
-from torch import nn
+from mmcv.cnn import ConvModule
+from torch import nn as nn
+from torch.utils import checkpoint as cp
 
 
 class InvertedResidual(nn.Module):
-    """Inverted residual module.
+    """InvertedResidual block for MobileNetV2.
 
     Args:
         in_channels (int): The input channels of the InvertedResidual block.
         out_channels (int): The output channels of the InvertedResidual block.
         stride (int): Stride of the middle (first) 3x3 convolution.
-        expand_ratio (int): adjusts number of channels of the hidden layer
+        expand_ratio (int): Adjusts number of channels of the hidden layer
             in InvertedResidual by this amount.
+        dilation (int): Dilation rate of depthwise conv. Default: 1
         conv_cfg (dict): Config dict for convolution layer.
             Default: None, which means using conv2d.
         norm_cfg (dict): Config dict for normalization layer.
             Default: dict(type='BN').
         act_cfg (dict): Config dict for activation layer.
             Default: dict(type='ReLU6').
+        with_cp (bool): Use checkpoint or not. Using checkpoint will save some
+            memory while slowing down the training speed. Default: False.
+
+    Returns:
+        Tensor: The output tensor
     """
 
     def __init__(self,
@@ -27,47 +34,59 @@ class InvertedResidual(nn.Module):
                  dilation=1,
                  conv_cfg=None,
                  norm_cfg=dict(type='BN'),
-                 act_cfg=dict(type='ReLU6')):
+                 act_cfg=dict(type='ReLU6'),
+                 with_cp=False):
         super(InvertedResidual, self).__init__()
         self.stride = stride
-        assert stride in [1, 2]
-
+        assert stride in [1, 2], f'stride must in [1, 2]. ' \
+            f'But received {stride}.'
+        self.with_cp = with_cp
+        self.use_res_connect = self.stride == 1 and in_channels == out_channels
         hidden_dim = int(round(in_channels * expand_ratio))
-        self.use_res_connect = self.stride == 1 \
-            and in_channels == out_channels
 
         layers = []
         if expand_ratio != 1:
-            # pw
             layers.append(
                 ConvModule(
-                    in_channels,
-                    hidden_dim,
+                    in_channels=in_channels,
+                    out_channels=hidden_dim,
                     kernel_size=1,
                     conv_cfg=conv_cfg,
                     norm_cfg=norm_cfg,
                     act_cfg=act_cfg))
         layers.extend([
-            # dw
             ConvModule(
-                hidden_dim,
-                hidden_dim,
+                in_channels=hidden_dim,
+                out_channels=hidden_dim,
                 kernel_size=3,
-                padding=dilation,
                 stride=stride,
+                padding=dilation,
                 dilation=dilation,
                 groups=hidden_dim,
                 conv_cfg=conv_cfg,
                 norm_cfg=norm_cfg,
                 act_cfg=act_cfg),
-            # pw-linear
-            nn.Conv2d(hidden_dim, out_channels, 1, 1, 0, bias=False),
-            build_norm_layer(norm_cfg, out_channels)[1],
+            ConvModule(
+                in_channels=hidden_dim,
+                out_channels=out_channels,
+                kernel_size=1,
+                conv_cfg=conv_cfg,
+                norm_cfg=norm_cfg,
+                act_cfg=None)
         ])
         self.conv = nn.Sequential(*layers)
 
     def forward(self, x):
-        if self.use_res_connect:
-            return x + self.conv(x)
+
+        def _inner_forward(x):
+            if self.use_res_connect:
+                return x + self.conv(x)
+            else:
+                return self.conv(x)
+
+        if self.with_cp and x.requires_grad:
+            out = cp.checkpoint(_inner_forward, x)
         else:
-            return self.conv(x)
+            out = _inner_forward(x)
+
+        return out
