+++ mmseg/utils/__init__.py
@@ -1,5 +1,4 @@
 from .collect_env import collect_env
-from .inverted_residual_module import InvertedResidual
 from .logger import get_root_logger
 
-__all__ = ['get_root_logger', 'collect_env', 'InvertedResidual']
+__all__ = ['get_root_logger', 'collect_env']
